fashion-search-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
weaviate-client==4.4.0
openai==1.3.5
requests==2.31.0
python-dotenv==1.0.0
watchdog==3.0.0
PyYAML==6.0.1
pillow==10.1.0
numpy==1.24.3
PyJWT==2.8.0
httpx==0.26.0
selectolax==0.3.17
aiohttp==3.9.1